#1 refactor ApprovedEntrant: hu.sch.domain.rest csomagba mozgatva, import fixek
