+++ sch-pek-domain/src/main/java/hu/sch/domain/rest/ApprovedEntrant.java
@@ -1,5 +1,6 @@
-package hu.sch.domain;
+package hu.sch.domain.rest;
 
+import hu.sch.domain.EntrantType;
 import java.io.Serializable;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
